requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-WTF==0.15.1
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
msgpack==1.0.4
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.19.1
requests==2.28.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==2.0.2
WTForms==2.3.3
